requirements.txt
numpy
panda
matplotlib
seaborn
scikit-learn
scipy

# pytorch packages
torch
torchvision
torchaudio